Input is now accepted by the program.
Input is an char** pointer, each string should be an integer with only one sign character in front. If a non-digit character is found, the program exits with with an error code. For this I made a very simple exit handler with an enum for the exit codes.

To free everything before the program ends, A new struct stacks is made and passed as argument so it can be freed in the exit handler when an error occures.