Very simple algorithm added.
Rotate smallest to the top, push to b stack. repeat until a is empty. Then since b is sorted in reverse, just push everything to the a stack.

I did this fairly quickly and the result is that it is very unoptimized. It does show me that it works, the checker as well, and that I would like some helper data in the structure and rewrite some functions.