Added general push function.
Push to a destination stack from a source stack.

At the moment it just uses ft_stack_front and ft_stack_del_front. Just changing where the pointers point to was a bit too long for the norm in my first implementation. Maybe if I don't use a member last in the stack.