Added ft_solve3 solving 3 numbers in max 2 operations.
This is done with the added function ft_check_ordered to check if, starting from the smallest value in the stack, looping once the values are ordered. This means the stack can be solved with just rotations. If this isn't the case, you just need one swap to make it so that it is (in case of just three numbers).