First stage pushes to b in groups, presorting and making fase two faster needing less rotations.
The amount of groups does have an impact. Sometimes when the count is over the required limit, changing the number of groups results in a count under the limit. After implementing simultainious rotations, maybe look for a way to decide the number of groups. Either some statistics or by doing it for multiple group counts and selecting the best one.